% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Load NOAA significant earthquake data into a tibble}
\usage{
eq_clean_data(NOAA_url)
}
\arguments{
\item{NOAA_url}{A character string giving the URL path to the significant
earthquake data file}
}
\value{
This function returns a tibble (data.frame) of the input file
}
\description{
This function reads tab-delimited file of the entire significant earthquake
data available at \code{NOAA_url} from U.S. National Oceanographic
and Atmospheric Administration (NOAA) database and returns a tibble (data.frame).
After reading the columns YEAR, MONTH and DAY are merged into column named
DATE, which contains Date class entries.
}
\note{
This function will give error if the file path is incorrect or
file does not exist
}
\examples{
NOAA_clean_data = eq_clean_data("https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt")

}
