% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Polish LOCATION_NAME column in the significant earthquake data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A tibble data.frame with LOCATION_NAME column that has to be cleaned}
}
\value{
This function returns a tibble (data.frame)
}
\description{
This function takes as input a data.frame of the entire significant
earthquake data \code{data}, cleans the LOCATION_NAME column by
stripping out the country name (including the colon) and converts
names to title case (as opposed to all caps).
}
\examples{
NOAA_url = "https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt"
data <- suppressMessages({readr::read_delim(NOAA_url, delim = "\\t", progress = FALSE)})
eq_location_clean(data)

}
