% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{filter_date_and_country}
\alias{filter_date_and_country}
\title{Earthquake data filtering}
\usage{
filter_date_and_country(data, xmindate, xmaxdate, country)
}
\arguments{
\item{data}{A preproccessed tibble (data.frame) of NOAA earthquakes
dataset}

\item{xmindate}{a character vector indicating the mininum date specified for
earthquakes time line plotting. Should follow the format of YYYY-MM-DD.}

\item{xmaxdate}{a character vector indicating the maximum date specified
for earthquakes time line plotting. Should follow the format of YYYY-MM-DD.}

\item{country}{a vector of character strings for the countries of interest.
By default it shows earthquake data related to USA.}
}
\value{
This function returns a tibble (data.frame)
}
\description{
This function filters the earthquake data \code{data} -
generated by U.S. National Oceanographic and Atmospheric Administration
(NOAA) database - based on specified minimum date \code{xmindate},
maximum date \code{xmaxdate} and country \code{country}.
}
\examples{
NOAA_clean_data = eq_clean_data("https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt")
NOAA_filtered = filter_date_and_country(NOAA_clean_data, xmindate = "1999-01-01", xmaxdate = "2017-12-31", country = c("CHINA", "USA"))
head(NOAA_filtered)

}
