% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Geom for plotting a time line of earthquakes with annotations}
\usage{
geom_timeline_label(data, xmindate, xmaxdate, country = c("USA"), n_max)
}
\arguments{
\item{data}{A preproccessed tibble (data.frame) of NOAA earthquakes
dataset}

\item{xmindate}{a character vector indicating the mininum date specified for
earthquakes time line plotting. Should follow the format of YYYY-MM-DD.}

\item{xmaxdate}{a character vector indicating the maximum date specified
for earthquakes time line plotting. Should follow the format of YYYY-MM-DD.}

\item{country}{a vector of character strings for the countries of interest.
By default it shows earthquake data related to USA.}

\item{n_max}{a numeric value for maximum number of earthquakes to be cosidered
for location annotations}
}
\value{
This function returns a ggplot
}
\description{
This function plots a time line of earthquakes from \code{data}
U.S. National Oceanographic and Atmospheric Administration
(NOAA) database. This geom adds a vertical line to each data point with
a text annotation (e.g. the location of the earthquake) attached to each line.
}
\examples{
NOAA_clean_data = eq_clean_data("https://www.ngdc.noaa.gov/nndc/struts/results?type_0=Exact&query_0=$ID&t=101650&s=13&d=189&dfn=signif.txt")
geom_timeline_label(NOAA_clean_data, xmindate = "1999-01-01", xmaxdate = "2017-12-31", country = c("CHINA", "USA"), n_max = 7.2)

}
